+++ src/gcovr/options.py
@@ -235,7 +235,7 @@ class Options:
 
 
 class GcovrConfigOptionAction(argparse.Action):  # pylint: disable=abstract-method
-    """Abstract class to be detect our own actions."""
+    """Abstract class for our own actions."""
 
     @abstractmethod
     def store_config_key(
@@ -270,12 +270,14 @@ class GcovrDeprecatedConfigOptionAction(GcovrConfigOptionAction):
         self, namespace: dict[str, Any], values: Any, config: str | None
     ) -> None:
         if config is not None:
+            values = self.value
             LOGGER.warning(
                 "Deprecated config key %s used, please use '%s=%s' instead.",
                 config,
                 self.config,
-                self.value,
+                values,
             )
+
         namespace[self.dest] = values