+++ CHANGELOG.rst
@@ -13,6 +13,10 @@ Breaking changes:
 - Jinja2 environment for HTML report is now created with ``undefined=StrictUndefined`` to raise an
   error if a variable is not defined in the template. (:issue:`1282`, :issue:`1283`)
 - Links to lines in HTML reports now use ``L<line>`` instead of a ``l<line>``. (:issue:`1285`)
+- Rename config key ``txt-metric`` (string) to ``txt-metrics`` (list of strings in TOML and multiple
+  entries in gcovr.config). Remove the deprecated aliases ``-b``, ``--txt-branches`` and
+  ``--branches``. (:issue:`1302`)
+- Adjust column width of total count in text reports. (:issue:`1302`)
 
 New features and notable changes:
 
@@ -29,6 +33,7 @@ New features and notable changes:
 - Add support for TOML files with :option:`--config` (:issue:`1258`)
 - Add :option:`--fail-under-condition-or-decision` with :option:`--fail-under-condition` and
   :option:`--fail-under-decision` as synonyms, change the key for the configuration file. (:issue:`1293`)
+- Add support for multiple metrics in one text report and add support for condition coverage. (:issue:`1302`)
 
 Bug fixes and small improvements:
 
@@ -581,7 +586,7 @@ New features and notable changes:
     (see `str.casefold <https://docs.python.org/3.11/library/stdtypes.html?highlight=str%20casefold#str.casefold>`_)
     (``file_10.c`` comes after ``file_0.c``).
   - Always sort at the end by filename if line or branch coverage is identical for a file.
-  - Add :option:`--sort-branches` to sort by branches instead of lines, this is the default if :option:`--txt-branches` is used.
+  - Add :option:`--sort-branches` to sort by branches instead of lines, this is the default if ``--txt-branches`` is used.
   - Add :option:`--sort-reverse` to reverse the sort order.
 
 - Add option to report covered lines in txt report. (:issue:`836`)
@@ -593,7 +598,7 @@ New features and notable changes:
 - Ignore all negative hits if :option:`--gcov-ignore-parse-errors` is used. (:issue:`852`)
 - Use literal options for sorting and TXT metric. (:issue:`867`)
 
-  - The :option:`-b`, :option:`--txt-branches` and :option:`--branches` are deprecated, use :option:`--txt-metric` instead.
+  - The ``-b``, ``--txt-branches`` and ``--branches`` are deprecated, use :option:`--txt-metric` instead.
     The reason for this is that we have line, branch and decision coverage and handle this with flags is more complex than
     using an enumeration.
   - The :option:`--sort-uncovered` and :option:`--sort-percentage` are deprecated, use :option:`--sort` instead.